% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_design_tibble.R
\name{as_design_tibble}
\alias{as_design_tibble}
\title{Transform data to tabular format}
\usage{
as_design_tibble(rigma_resp, message = TRUE)
}
\arguments{
\item{rigma_resp}{A response object from Figma API.}

\item{message}{logical. Control printing of messages.}
}
\value{
tibble data extracted from Figma response objects. Subclasses
of type `design_tibble` are added to the resulting tibbles.
}
\description{
Transforms data returned from a Figma API request to tabular
format.
}
\examples{
\dontrun{
file_key <- "sFHgQh9dL6369o5wrZHmdR"
resp <- get_file_styles(file_key = file_key)
as_design_tibble(resp)
}

}
