% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments}
\alias{get_comments}
\title{GET comments}
\usage{
get_comments(file_key)
}
\arguments{
\item{file_key}{string. The key that a Figma file is referred by.}
}
\description{
Gets a list of comments left on the file.
}
\examples{
\dontshow{if (Sys.getenv("FIGMA_ACCESS_TOKEN") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
#navigate to  file and get key from url
file_key <- "sFHgQh9dL6369o5wrZHmdR"
get_comments(file_key)
}
\dontshow{\}) # examplesIf}
}
