% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_component_sets.R
\name{get_file_component_sets}
\alias{get_file_component_sets}
\title{GET file component sets}
\usage{
get_file_component_sets(file_key)
}
\arguments{
\item{file_key}{string. Id of the team to list components from.}
}
\value{
S3 object of class \code{rigma_get_file_component_sets}. Components are
stored in the \code{meta} field.
}
\description{
Get a list of published component_sets within a file library.
Note that published components are only available via the professional
plan.
}
\examples{
\dontshow{if (Sys.getenv("FIGMA_ACCESS_TOKEN") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
#navigate to team page and get id from url
file_key <- "gYRjH0y8ZM0VtEfO8kf6ch"
get_file_component_sets(file_key)
}
\dontshow{\}) # examplesIf}
}
