% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_styles.R
\name{get_file_styles}
\alias{get_file_styles}
\title{GET file styles}
\usage{
get_file_styles(file_key)
}
\arguments{
\item{file_key}{string. The key that a Figma file is referred by.}
}
\value{
S3 object of class \code{rigma_get_file_styles}. Styles are
stored in the \code{meta} field.
}
\description{
Get published styles in a file library by name.
}
\examples{
\dontshow{if (Sys.getenv("FIGMA_ACCESS_TOKEN") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
#navigate to team page and get id from url
file_key <- "gYRjH0y8ZM0VtEfO8kf6ch"
get_file_styles(file_key)
}
\dontshow{\}) # examplesIf}
}
