% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_versions.R
\name{get_file_versions}
\alias{get_file_versions}
\title{GET file versions}
\usage{
get_file_versions(file_key)
}
\arguments{
\item{file_key}{string. The key that a Figma file is referred by.}
}
\value{
S3 object of class `rigma_get_file_versions`. Contains the parsed
JSON response with fields `versions`, and `pagination`.
}
\description{
A list of the versions of a file.
}
\examples{
\dontshow{if (Sys.getenv("FIGMA_ACCESS_TOKEN") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
#navigate to  file and get key from url
file_key <- "sFHgQh9dL6369o5wrZHmdR"
get_file_versions(file_key)
}
\dontshow{\}) # examplesIf}
}
