% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_files.R
\name{get_project_files}
\alias{get_project_files}
\title{GET project files}
\usage{
get_project_files(project_id, branch_data = "false")
}
\arguments{
\item{project_id}{string. Id of the project to list files from.}

\item{branch_data}{string. Returns branch metadata in the response for each
main file with a branch inside the project. Default: "false"}
}
\description{
List the files in a given project.
}
\examples{
\dontshow{if (Sys.getenv("FIGMA_ACCESS_TOKEN") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
#navigate to  file and get key from url
project_id <- "71686204"
get_project_files(project_id)
}
\dontshow{\}) # examplesIf}
}
