% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{possibly_thumbnail_color}
\alias{possibly_thumbnail_color}
\title{Possibly extract thumbnail color}
\usage{
possibly_thumbnail_color(...)
}
\value{
thumbnail_color function that handles errors
}
\description{
Possibly extract thumbnail color
}
