% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_figma.R
\name{req_figma_template}
\alias{req_figma_template}
\title{Set the API URL template based on the specified endpoint}
\usage{
req_figma_template(
  req,
  endpoint = c("file", "file nodes", "images", "image fills", "versions", "users",
    "comments", "comment reactions", "projects", "project files", "team components",
    "file components", "component", "team component sets", "file component set",
    "component set", "team styles", "file styles", "style"),
  template = NULL,
  ...,
  call = rlang::caller_env()
)
}
\arguments{
\item{endpoint}{Figma REST API end point to use}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Set the API URL template based on the specified endpoint
}
\keyword{internal}
