% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_rigma_agent.R
\name{req_rigma_agent}
\alias{req_rigma_agent}
\title{Add metadata to Rigma request}
\usage{
req_rigma_agent(req, user_agent = "Rigma https://github.com/AleKoure/Rigma")
}
\arguments{
\item{req}{httr2_request object}

\item{user_agent}{string with user agent}
}
\value{
httr2_request object with Figma token heater, user agent and retry
specifications.
}
\description{
Extra metadata attached to httr2 request object before
performing the request.
}
