% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_figma.R
\name{request_figma_endpoint}
\alias{request_figma_endpoint}
\title{Create a request for a specified Figma API Endpoint}
\usage{
request_figma_endpoint(
  endpoint = NULL,
  base_url = "https://api.figma.com",
  template = NULL,
  method = NULL,
  token = NULL,
  ...
)
}
\arguments{
\item{endpoint}{Figma REST API end point to use}

\item{base_url}{Base URL for request.}

\item{method}{Custom HTTP method}

\item{token}{Figma API access token. If `NULL`, replace with environment variable named "FIGMA_ACCESS_TOKEN" or an alternate name set using the variable "Rigma.figma_token".}

\item{...}{
  Arguments passed on to \code{\link[httr2:req_template]{httr2::req_template}}
  \describe{
    \item{\code{req}}{A \link[httr2]{request}.}
    \item{\code{template}}{A template string which consists of a optional HTTP method
and a path containing variables labelled like either \verb{:foo} or \code{{foo}}.}
    \item{\code{.env}}{Environment in which to look for template variables not found
in \code{...}. Expert use only.}
  }}
}
\description{
Create a request for a specified Figma API Endpoint
}
\keyword{internal}
