% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rigma-package.R
\docType{package}
\name{Rigma-package}
\alias{Rigma}
\alias{Rigma-package}
\title{Rigma: Access to the 'Figma' API}
\description{
The goal of Rigma is to provide a user friendly client to the 'Figma' API \url{https://www.figma.com/developers/api}. It uses the latest `httr2` for a stable interface with the REST API. More than 20 methods are provided to interact with 'Figma' files, and teams. Get design data into R by reading published components and styles, converting and downloading images, getting access to the full 'Figma' file as a hierarchical data structure, and much more. Enhance your creativity and streamline the application development by automating the extraction, transformation, and loading of design data to your applications and 'HTML' documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AleKoure/Rigma}
}

}
\author{
\strong{Maintainer}: Alexandros Kouretsis \email{alexandros@appsilon.com}

Authors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID})
}

}
\keyword{internal}
