% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figma_token.R
\name{set_figma_token}
\alias{set_figma_token}
\alias{get_figma_token}
\title{Set or get a Figma personal access token (PAT) saved as an environment
variable}
\usage{
set_figma_token(
  token,
  install = FALSE,
  overwrite = FALSE,
  quiet = FALSE,
  call = caller_env()
)

get_figma_token(token = NULL, call = caller_env())
}
\arguments{
\item{token}{A Figma personal access token (PAT). Optional for
\code{\link[=get_figma_token]{get_figma_token()}}.}

\item{install}{If \code{TRUE}, this function adds your token to your \code{.Renviron}
for use in future sessions. Defaults to \code{FALSE}.}

\item{overwrite}{If \code{TRUE}, overwrite any existing token in \code{.Renviron} using
the same environment variable name. Defaults to \code{FALSE}.}

\item{quiet}{If \code{TRUE}, suppress messages when setting token by locally
setting the \code{cli.default_handler} option to \code{\link[=suppressMessages]{suppressMessages()}}. Defaults
to \code{FALSE}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A string with a Figma personal access token.
}
\description{
\code{\link[=set_figma_token]{set_figma_token()}} can set personal access token (PAT) as a local
environment variable temporarily for the current session or save the token
for future sessions.

\code{\link[=get_figma_token]{get_figma_token()}} returns a saved Figma PAT or errors if the token is
missing or if the token does not match the expected pattern.
}
