% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumbnail_color.R
\name{thumbnail_color}
\alias{thumbnail_color}
\title{Find thumbnail color}
\usage{
thumbnail_color(path, hex = TRUE)
}
\arguments{
\item{path}{string. Path to the thumbnail PNG}

\item{hex}{logical. If \code{TRUE} then the RGBA values are converted to hex}
}
\value{
The color of the thumbnail in hex or rgba
}
\description{
Find thumbnail color
}
\details{
Given a mono-colored thumbnail this function extracts the RGBA
channels and returns a vector scaled from \link{0, 1} or a hex color code.
}
\examples{
path <- system.file("extdata", "test_thumbnail.png", package = "Rigma")
thumbnail_color(path)

}
