% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_component_sets.R
\name{get_component_sets}
\alias{get_component_sets}
\title{GET component set}
\usage{
get_component_sets(key)
}
\arguments{
\item{key}{string. The unique identifier of the component.}
}
\value{
S3 object of class \code{rigma_get_component_set}.
}
\description{
Get metadata on a component_set by key. To publish components
join a Figma team and subscribe for professional account.
}
\examples{
\dontrun{
component_key <- "my_key"
get_component_sets(component_key)
}

}
