% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_figma_template.R
\name{req_figma_template}
\alias{req_figma_template}
\title{Set the API URL template based on the specified endpoint}
\usage{
req_figma_template(
  .req,
  endpoint = c("file", "file nodes", "images", "image fills", "versions", "users",
    "comments", "comment reactions", "projects", "project files", "team components",
    "file components", "component", "team component sets", "file component set",
    "component set", "team styles", "file styles", "style"),
  .perform = FALSE,
  ...,
  call = caller_env()
)
}
\arguments{
\item{.req}{A \link[httr2]{request}.}

\item{endpoint}{Figma REST API endpoint to use}

\item{...}{Template variables.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Modify a request using a template for a specified Figma API endpoint.
}
\keyword{internal}
