% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_figma.R
\name{request_figma}
\alias{request_figma}
\title{Create a request for the Figma API with an optional method}
\usage{
request_figma(base_url = "https://api.figma.com", method = NULL, token = NULL)
}
\arguments{
\item{method}{Custom HTTP method}

\item{token}{Figma API access token. If \code{NULL}, replace with environment
variable named "FIGMA_ACCESS_TOKEN" or an alternate name set using the
variable "Rigma.figma_token".}

\item{name}{Base URL for request. Defaults to \code{"https://api.figma.com"}}

\item{...}{
  Arguments passed on to \code{\link[httr2:req_template]{httr2::req_template}}
  \describe{
    \item{\code{req}}{A \link[httr2]{request}.}
    \item{\code{template}}{A template string which consists of a optional HTTP method
and a path containing variables labelled like either \verb{:foo} or \code{{foo}}.}
    \item{\code{.env}}{Environment in which to look for template variables not found
in \code{...}. Expert use only.}
  }}
}
\description{
Use httr2::request() to create a request with error handling an optionally
specified method.
}
\keyword{internal}
